added latest INQ implementation incl. uniform (i.e. non-power-of-2) levels incl. bug-fix added latest RPR implementation within INQ module modified STE to support monitoring epoch
